refactor(workspace): merge `running-process-client` into `running-process` (Wave 4/9 of #165)
Absorbs the IPC client crate (lib + `runpm` PM2-style CLI binary) into
the consolidated mono-crate. Drops the workspace from 7 → 6 members.

## Cargo.toml changes
- `[features] default = ["client"]` (per Q1 resolution in #165 — needed
  so `cargo install running-process` installs `runpm` out of the box;
  `required-features` are NOT auto-activated by cargo install per
  rust-lang/cargo#4663).
- `client = ["dep:prost", "dep:prost-types", "dep:interprocess", "dep:dirs", "dep:anyhow", "dep:clap"]`
  — IPC dep set is now feature-gated. Wave 3 had added prost/prost-types
  unconditionally; this wave moves them behind `client` along with the
  rest of the IPC deps as planned.
- `[[bin]] runpm` with `required-features = ["client"]`.

## Source moves (via git mv, blame preserved)
- `crates/running-process-client/src/{client,paths,pipe_session,pty_session}.rs`
  → `crates/running-process/src/client/`
- New `client/mod.rs` re-exports the public API at `running_process::client::*`
- `crates/running-process-client/src/bin/runpm.rs`
  → `crates/running-process/src/bin/runpm.rs`

## lib.rs
- `pub mod proto` is now `#[cfg(feature = "client")]`-gated alongside
  `pub mod client` (Wave 3 had left it unconditional pending this wave).

## Consumer updates
- `running-process-daemon` Cargo.toml: drop `running-process-client` dep
  (now reachable via `running-process` with default `client` feature
  active). Source: `running_process_client` → `running_process::client`
  sed across 4 files.
- `running-process-py` Cargo.toml: drop `running-process-client` dep,
  add `features = ["client", "originator-scan"]` to its
  `running-process` dep (explicit even though defaults would suffice).
  Source: same sed across 2 files.

`cargo check --workspace` clean. Workspace member count: 6.

Closes #169. Ship-and-merge per #165.
