+++ Cargo.toml
@@ -2,7 +2,6 @@
 resolver = "2"
 members = [
     "crates/running-process",
-    "crates/running-process-client",
     "crates/running-process-daemon",
     "crates/daemon-trampoline",
     "crates/running-process-py",
