--- crates/running-process-client/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "running-process-client"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-description = "Lightweight synchronous IPC client for the running-process daemon"
-
-[[bin]]
-name = "runpm"
-path = "src/bin/runpm.rs"
-
-[dependencies]
-# Wave 3 of #165: proto types now live in `running-process::proto`
-# (the former `running-process-proto` crate was merged in).
-running-process = { path = "../running-process" }
-prost = "0.14"
-interprocess = "2"
-dirs = "6"
-# Used by the `runpm` PM2-style supervisor CLI (src/bin/runpm.rs).
-anyhow = "1"
-clap = { version = "4", features = ["derive"] }
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
