+++ crates/running-process-daemon/Cargo.toml
@@ -16,7 +16,6 @@ path = "src/main.rs"
 
 [dependencies]
 running-process = { path = "../running-process", version = "3.4.1" }
-running-process-client = { path = "../running-process-client", version = "3.4.1" }
 prost = "0.14"
 tokio = { version = "1", features = ["full"] }
 tokio-util = { version = "0.7", features = ["codec"] }
