+++ crates/running-process-py/Cargo.toml
@@ -16,9 +16,8 @@ crate-type = ["cdylib", "lib"]
 libc = "0.2"
 pyo3 = { workspace = true }
 regex = "1"
-running-process = { path = "../running-process", version = "3.4.1", features = ["originator-scan"] }
+running-process = { path = "../running-process", version = "3.4.1", features = ["client", "originator-scan"] }
 sysinfo = "0.30"
 winapi = { version = "0.3", features = ["consoleapi", "handleapi", "jobapi2", "processenv", "processthreadsapi", "synchapi", "winbase", "wincon", "winnt", "winuser"] }
-running-process-client = { path = "../running-process-client", version = "3.4.1" }
 prost = "0.14"
 interprocess = "2"
