+++ crates/running-process/Cargo.toml
@@ -8,21 +8,26 @@ repository.workspace = true
 homepage.workspace = true
 description = "Subprocess and PTY runtime for the running-process project"
 
+[[bin]]
+# Wave 4 of #165: PM2-style supervisor CLI absorbed from
+# `running-process-client`. Requires the `client` feature for IPC deps.
+name = "runpm"
+path = "src/bin/runpm.rs"
+required-features = ["client"]
+
 [features]
-# Wave 2/9 of #165: feature scaffolding for the upcoming mono-crate
-# consolidation. Subsequent waves (W3-W6) merge the proto / client /
-# daemon / trampoline crates into this one and switch these features
-# on. Final feature scheme is documented in #165.
-#
-# `default` will become `["client"]` once Wave 4 lands and populates
-# the `client` feature with its actual dep list (per Q1 resolution in
-# #165). For now `default` stays `[]` so this wave is a pure no-op
-# diff at the dep-graph level.
-default = []
-core = []                       # always-available API; placeholder marker
-client = []                     # populated by Wave 4 (#169)
-daemon = ["client"]             # populated by Wave 5 (#170)
-originator-scan = []            # used by running-process-py for cwd-tagging
+# Final feature scheme per #165:
+# * `core`   — always-available API (spawn / pty / containment).
+# * `client` — adds proto types + IPC client (prost, interprocess, dirs).
+# * `daemon` — superset of client; adds the full daemon runtime (Wave 5).
+# Default ships `client` so `cargo install running-process` installs
+# the `runpm` binary out of the box (per Q1 resolution in #165:
+# `required-features` are not auto-activated by cargo install).
+default = ["client"]
+core = []
+client = ["dep:prost", "dep:prost-types", "dep:interprocess", "dep:dirs", "dep:anyhow", "dep:clap"]
+daemon = ["client"]                # populated by Wave 5 (#170)
+originator-scan = []               # used by running-process-py for cwd-tagging
 
 [dependencies]
 libc = "0.2"
@@ -30,12 +35,15 @@ portable-pty = "0.9"
 sysinfo = "0.30"
 thiserror = { workspace = true }
 winapi = { version = "0.3", features = ["errhandlingapi", "fileapi", "handleapi", "ioapiset", "jobapi2", "namedpipeapi", "processthreadsapi", "winnt", "minwindef", "windef", "winuser", "consoleapi", "processenv", "synchapi", "winbase", "wincon", "tlhelp32"] }
-# Wave 3 of #165: proto types absorbed from the former
-# `running-process-proto` crate. Unconditional today; Wave 4 will
-# move these behind `feature = "client"` along with the rest of the
-# IPC subsystem.
-prost = "0.14"
-prost-types = "0.14"
+# Wave 4 of #165: client-feature deps. All optional so the always-on
+# `core` API stays a minimal-dep leaf.
+prost = { version = "0.14", optional = true }
+prost-types = { version = "0.14", optional = true }
+interprocess = { version = "2", optional = true }
+dirs = { version = "6", optional = true }
+# Used by the `runpm` PM2-style supervisor CLI (src/bin/runpm.rs).
+anyhow = { version = "1", optional = true }
+clap = { version = "4", features = ["derive"], optional = true }
 
 [build-dependencies]
 # Wave 3 of #165: build.rs absorbed from `running-process-proto`.
